Support mime encoding current subtree
This doesn't require that the user explicitly narrow to a subtree, and
it adds support for export options set as properties on the subtree.

The latter is fairly crucial as it lets one do things such as suppress
table of contents generation on a subtree without doing so for the
entire document.