Fix file attachments and equations and enhance emails from org-files
Major changes:
1. I think that file links should be attachments in an email. This commit makes
that happen. It should work in all three ways of making/sending messages. This
only works for org 9.0, and it is conditional. older versions of org just won't
attach files.
2. Add file keyword support for email headers when using
org-mime-org-buffer-htmlize, and update docstrings. This is also a conditional,
org 9 feature.
3. Equations were not getting correctly handled in org-mime-org-buffer-htmlize
and org-mime-org-subtree-htmlize. The images were not getting created and
attached correctly. This commit fixes that.
4. org-mime--chomp is too aggressive, and does more than the docstring suggests.
I changed the docstring, and removed the call to it in org-mime-htmlize. Now the
function is not used anywhere, so it could be removed. It would be preferrable
to use something like s-trim.
5. It should be possible to use relative paths to images with this patch.
6. Add an org export to org-mime-htmlize. Previously no export was done for
the plain mimepart. Some links don't make sense in the email then, so adding the
org export allows them to be re-rendered in a way that does make sense. The
most concrete example is a cite/bibliography link from org-ref, which now get
converted to ascii representations when the email is htmlized.

Minor fixes:
1. org-export-htmlize-output-type is no longer a variable in org-9. It is now
org-html-htmlize-output-type. Both variables are still used for backward
compatibility.
2. a few typos
3. Some comment clarifications