Fix file attachments and equations and enhance emails from org-files
Thanks to John Kitchin (AKA jkitchin) for this patch.

See https://github.com/org-mime/org-mime/pull/20

Major changes:

- I think that file links should be attachments in an email. This commit makes
that happen. It should work in all three ways of making/sending messages.

- Add file keyword support for email headers when using
org-mime-org-buffer-htmlize, and update docstrings.

- Equations were not getting correctly handled in org-mime-org-buffer-htmlize
and org-mime-org-subtree-htmlize. The images were not getting created and
attached correctly. This commit fixes that.

- org-mim--chomp is too aggressive, and does more than the docstring suggests.
I changed the docstring, and removed the call to it in org-mime-htmlize. Now the
function is not used anywhere, so it could be removed. It would be preferrable
to use something like s-trim.

- It should be possible to use relative paths to images.

- Add an ascii export to org-mime-htmlize. Previously no export was done for
the plain mimepart. Some links don't make sense in the email then, so adding the
ascii export allows them to be re-rendered in a way that does make sense. The
most concrete example is a cite/bibliography link from org-ref, which now get
converted to ascii representations when the email is htmlized.

Minor fixes:

- org-export-htmlize-output-type is no longer a variable in org-9. It is now
org-html-htmlize-output-type. Both variables are still used for backward
compatibility.

- a few typos

- Some comment clarifications