Show individual paragraphs in gmail reply quotes
Prior to this patch, replies using the org-mime-beautify-quoted-mail option
would show up with multiple paragraphs in the reply squashed onto a single
line (at least in gmail).

For example, the reply:

```
> This is one para (which is
> split across two lines)
>
> This is another
```

Would get rendered as:

```
<p>
This is one para (which is
split across two lines)

This is another
</p>
```

And displayed in gmail as:

```
| This is one para (which is split across two lines) This is another
```

(Where '|' is the beautifed vertical line in gmail.)

Looking at how gmail itself deals with multiple paras in a reply, it appears
they:

1. put each paragraph in its own div
2. represent blank lines between divs as <br />

They do some funky nesting that doesn't seem to be needed to make the quotes
display correctly, so I've left it out.

This patch would render the example above into the structure:

```
<div>This is one para (which is
split across two lines)</div>
<div><br/></div>
<div>This is another</div>
```

(With some less beautiful line breaks) and displayed in gmail as:

```
| This is one para (which is split across two lines)
|
| This is another
```

To better represent the gmail de-factor standard, this patch also changes the
top-level element in the gmail blockquote to div (from p).