fix: Use const in the :type of `org-mime-library' and `org-mime-export-ascii'
When using the choice constructor in the :type of defcustoms, symbols
are denoted as the value of a const constructor (see (elisp) Composite
Types).  Update the defcustoms of `org-mime-library' and
`org-mime-export-ascii' such that they obey this syntax.

Without this fix, using the Emacs 29+ setopt form will error, since
the type of these options were not indicated properly.

Additionally, make the docstrings of these user options more
informative.