fix: Saving in the org-mime org buffer no longer duplicates text
Currently, in `org-mime-edit-src-save`, the region between
`org-mime-src--beg-marker` and `org-mime-src--end-marker` would be
deleted followed by a replacement of the text in the org-mime org
buffer.

However, the deletion of that text would not move
`org-mime-src--end-marker` as desired: to the end of the new email
body.  Consequently, subsequent saves in the org-mime org buffer could
duplicate text.

The problem arose because the TYPE argument of the `copy-marker` used
to create `org-mime-src--end-marker` was left nil.  Instead, a value
of t moves the marker to the end of text inserted at its
location.:(For an explanation of why, see (elisp) Marker Insertion
Types.)  Therefore, pass t to the TYPE argument of the `copy-marker`
used to create `org-mime-src--end-marker.